Fix inconsistency in how guards see around corners
Previously if they were in a north-south door they could only see the square in front of them, while if they were in an east-west door they could see the squares to the left and right as well (along with the diagonals going out from them). This is because the ray was passing exactly through the grid vertex and we weren't considering whether horizontal-then-vertical and vertical-then-horizontal stepping would produce blocking results.