Treasure vault rooms
These show up currently at level 5 and above. If there are dead rooms, the level generator picks one at random and makes it into a vault.

Vaults have extra treasure, a locked door (although it is not locked yet), and guards do not patrol into the vault (because it's locked). They also have no windows and a unique floor type.