Remove bias on which guards carry key and loot
Previously the patrol routes were ordered starting from the lower left, and we handed out the key to the first patrol route, loot to subsequent patrol routes. This meant the key-carrying guard was always in the lower left. Shuffle the patrol routes (before adding the special outside patrol routes).

Remove standing near loot as a thing for guards to do at the ends of patrol routes. This lets us generate the patrol routes before placing loot, so we know how many guards there will be and can avoid handing out more loot than there are guards to carry it.