Allow movement onto a moving-to-investigate guard again
The problem with this previously was that when investigating, guards check sight at the beginning of act(), so they can immediately go into ChaseVisibleTarget and move toward the player. In this case the player was directly on top of the guard, so the guard would not move and would be underneath. I had tried disallowing movement onto a moving-to-investigate guard but this prevents the player from being able to sneak up behind them to knock them out or pick their pocket.

Instead, allow movement onto the moving guard, but when doing the sight check at the top of act(), require the player to be in front of the guard to be seen (rather than in front of or even with the guard). This allows the guard to go ahead and take a step forward, preserving the appearance of simultaneous action.