Vault room-selection improved
Allow dead-end courtyard rooms to be converted into vaults. This used to not work due to the order in which walls got plotted, but things have been cleaned up so it works now.

If there are multiple candidate dead-end rooms, choose randomly from the ones with the smallest area, rather than choosing randomly from among all of them.