Compute betweenness centrality for rooms
Currently it is not used for anything, but could eventually be used as a better indicator of how high-traffic a room is. The current generator is using the number of room exits as a heuristic for how high-traffic it is, which is pretty good. Occasionally you have a room with only two exits that is on the required path from the entrance to the rest of the house, though.