Ensure placed furniture is all reachable
Use a bit more efficient of an approach, by identifying all usable furniture positions in one depth-first search, for each piece of furniture placed. Additionally, if any piece of furniture has only one walkable space next to it, mark that space as unusable to ensure the furniture will not get completely boxed in.

Allow bedrooms to be width or height 2 (but not both). Ensure the bed does not bisect the room if the room's width is the same as the bed's.