Don't merge non-coincident adjacencies between same rooms
This was a bug when a room stuck out on both the left and the right side of the building. In this situation there would be adjacencies between the room and the exterior room on the left and the right, and they were getting merged because they were collinear and had the same rooms on either side. This was a problem because it made an adjacency that crossed the interior part of the building.

Now it checks that two candidate adjacencies actually abut each other, in addition to being collinear and separating the same two rooms.