Add kitchen room type, stove and shelf item types
The kitchen decoration is really lame right now, as is the selection of a room to be the kitchen. This is mostly a placeholder to get something in.

The stove functions as a light source that you cannot leap over or extinguish. On the other hand it blocks guard sight.

The shelf functions identically to the bookshelf; it just looks empty.