Better kitchen decoration
The furniture is laid out galley-kitchen style, along two opposing walls. If there is a lot of room in the middle, additional tables are placed there.

The number of stoves scales with the size of the room.

Additionally, kitchen is prevented from being placed in a room with an exterior door, and from being placed in a house with less than eight rooms.