Very simple leap trainer
On level three (currently), start the player outside the mansion, just around the corner from a patrolling guard. Unless the player takes immediate action they will be spotted and have to get away, and running is the easiest way to do that outside the mansion.

Put up a prompt for the first 16 turns suggesting using leap.