Health pickups heal
Also place them a bit more carefully. There's a formula for the number of health pickups as a function of level. The first level has no health; the second has three; and then it drops off to one as the game goes on.

The first health pickup is always in the kitchen, if it's available. The next health pickup is always in either the kitchen or the dining room, if possible. After that they are in kitchen, dining, or bedrooms.