Slightly more granular ghost bonus
Collect counts of the number of times the player is spotted (double-red exclamation mark), number of times the player is damaged, and number of times the player knocks an enemy out. Add 1/3 to the score multiplier for each of them that is zero at the end of the level. So if you don't get spotted, don't knock anyone out, and don't get damaged, the multiplier is 2x. If you get spotted but don't get damaged or knock anyone out, the multiplier is 5/3. If you get spotted and injured but don't knock anyone out, the multiplier is 4/3. If all three happen the multiplier is 1.

Eventually I would like to collect a bunch more stats to present to the player for their amusement at the end of the game.