Force guard to stay in ChaseVisibleTarget for at least one turn
This keeps them from entering it at the beginning of a turn and then leaving it at the end of the same turn, and thus never appearing to have entered it, even though they start moving to the last sighting.

This makes the game slightly harder because the guards will shout when they begin moving to you, and you will lose your ghost status.