Tighten up guard sight when suspicious and stationary
Previously they were applying the look-one-step-ahead rule for the pre-act sighting, which is what lets the player step into the guard's vacated position. But if they are standing still this was causing them to take an additional turn to give chase, which felt inconsistent.

Now when a guard is not expecting to move they will use sight that can see to either side.