Restrict pathing from one door in a room to another door in a room to be only through that room
When generating patrol routes it is occasionally cheaper for a guard to get between two doors in a room by moving outside of the room (for instance, if the room is full of furniture and requires a circuitous route, while there is a hall outside connecting the doors directly).

Since the patrol-route generation is expecting the guards to move through the room between the two doors this resulted in very odd patrol routes.

Now it only generates a distance field for the room itself, which is cheaper too.