Make status bar fit in 46 characters
We still target 80 characters so there will be breathing room (and the font won't be too big), but on the iPhone SE the horizontal resolution only allows for about 46 characters. We don't want to scale the font down because it gets too tiny, so do some juggling on the status bar to tighten it up.