Remove a feature I accidentally added
When the on-screen gamepad was displayed, it panned the world so the corners were inside the gamepad buttons, so they wouldn't be obscured by the buttons. I've decided to just make the buttons more transparent and require the player to pan to see what's going on in the world corners, since the game does not go into the corners much.