Adjust scoring algorithm
There's a completion score which is 10x the loot count for the level. This is added to the time under par; the sum is then doubled if the level was Ghosted. This ensures there's always a non-zero score for completing the level.

The par time is a multiple of 10, to make it feel more rounded.