Trying out a different scoring algorithm
This is primarily based on time under par. The par tries to account for the amount of time it will take to deal with getting spotted by guards once or twice. There's a small fixed bonus of 20 points for ghosting the level.

Show the number of turns taken on the end-of-level screen, and show it on the status bar.