Guards plan their moves before the player moves
This allows us to prevent letting the player jump into a guard's current position, if it will involve jumping over the guard's intended next position. Previously we were able to do this for patrol paths but not for the other states, which could lead to some odd situations.

Additionally, guards will fall back to less-desirable moves if necessary, to avoid being on the same square as a guard who has already moved. Previously this worked if the guard was pathing toward a position, but did not work if the guard was trying to stand still to do something, because they didn't do any consideration of positions or movement.

Simplify the guard update order to only prioritize guards actively chasing the player. The distance-based sorting could cause oscillations sometimes.

Guards moving to investigate a noise will move to a square adjacent to the noise rather than to the noise position itself.