Show guard speech in speech bubbles
Hopefully this is not too obtrusive; it gets it closer to where the player's eye is. Speech can still go by quickly if it gets replaced by speech on a subsequent turn. The last speech can be recalled with the Tab key. I'd like to do this with a tap of the Jump key but that requires rewriting the controllers.

Fix some of the subtitles, and reorder speech priorities a bit so when someone hears and moves to investigate their speech is prioritized over someone who hears but won't investigate. Also prefer the closest speaker to the player for identical speech events.