Level-generation adjustments
Fortresses now have a U-shaped interior courtyard instead of a complete ring, half the time.
The single window on a wall of length four (measured in the room interior) will be offset so it's not always in the same place.
The room shapes are always mirrored horizontally (again), but room adjacencies are sometimes not mirrored.