Fortress level-generation changes
There are windows to the outside now, but only single windows on walls where there can be a single window in the center.

Windows to courtyards are only placed if the courtyard is at a lower depth than the room.

Adjacencies are not mirrored. The level shape is constrained enough already.

No back or side doors/gates.