Replace top status bar with a floating status message
This is more visible than the status bar since it is over the game area. Hopefully not too annoying.

It positions itself at the top or bottom of the screen depending on whether the player is in the bottom or top half of the map.

I originally tried basing it on the player's on-screen position but due to animation that was less stable.