Tutorial tuning
Get rid of the "sticky" feature of the status bar which was only used for the you-are-dead message.
Put the status-bar message computation into a function for better control flow.
Put arrow icons in the leap messages so it matches the player's attempted move direction.
Use a checkbox instead of a counter on the prompts that require multiple successes, to make it less confusing.
Hide the prompt about opening the menu after the first few turns, for players who already know what they're doing.
Remove the lit/unlit status messages.