Speech-bubble and pop-up fixes
Use guard's target position instead of animated position for deciding whether to have the speech bubble above or below the guard.

Use hysteresis for swapping speech-bubble sides; if the guard's Y position is the same as the player's Y position, the speech bubbble will not flip sides.

Clear the pop-up notification at the start of each turn. It does not make sense to carry over notifications from previous turns for recall using Tab, because they tend to be related to the player's last action.

When recalling a speech bubble via Tab, snap it to its correct side for the current guard/player vertical relationship.