Generic public room decoration work
Pools are oriented to be more square, with space between them and the pillars for the most part.
Add torches to pillar rooms.
Add chairs to pillar rooms, if they are big enough.

Dining rooms can be in public or private areas, not just in public areas.

Flooring in treasure room changes depending on whether it is in a public or private area of the house.

Stationary guards will not stand at a window with a lantern behind them.