Locked treasure fixes
When a treasure gets unlocked, only remove its lock, not all of the locks in the level.

A book switch can participate in multiple treasure locks. Each lock advances, completes, and resets independently. The sound effect is based on the highest-priority thing that happened as a result of pushing the switch.

Don't use the same books for different treasure locks, though. Shuffle all the books and then deal them out to switches.