Level-creation work
Fix more cases where doors could be created in walls of length one, which doesn't work.

Sort patrol routes by the minimum depth of the rooms they pass through. This puts loot guards in the interior of the building.

Put the key on the median route, so the key guard won't be encountered right at the front, but also not at the very back.