Show failed achievements
Achievements have been reworked a bit so they declare failure for the current run as early as possible.

On the achievements page, an incomplete achievement that is no longer achievable on the current run has a blank space instead of an empty checkbox. (This seemed like the easiest indicator to get working.)