Fix minor issue with guards stopping at activity stations
The algorithm for deciding whether to insert an extra turn now more accurately determines whether the guard will need to turn when they arrive at the station. Previously if they approached at a diagonal it wouldn't, which made it appear that they only spent two turns facing in the activity direction.