Move noise-making to a dedicated button
For now it's X, the top face button on the right side of the controller, and a bracket symbol in the touch controls. (I think this icon is unused and could be replaced with a noise-maker icon.)

The game tries to pick a direction to bang, to preserve the old bang-the-wall tactile feel. If you do it in open ground it goes downward which implies banging the ground.