Map tuning
Map dimensions are computed by summing up a bunch of d2's, so there's more of a normal distribution instead of a uniform distribution.

The manor algorithm for generating inside/outside areas now works slightly differently. It throws darts at the whole grid to make outside rooms, and then sometimes mirrors the left side to the right side, and/or the bottom to the top.

Fortresses can be a bit smaller.