More room-width variation in warrens
I want things to be pretty cramped, but if I squeeze the room size down there's not much space for larger room styles, or for more random offset variation.

Pick randomly from a base width of 6 or 7 for each row and column in the grid. Then offset by up to 2 or 3, so the minimum room size is 4.