Change algorithm for placing exterior doors
It's looking for edges that cross the centerline, and from those, whichever is farthest toward the edge of the map.

The probabilities for having back and side doors are guaranteed if the outside perimeter exceeds a threshold length. This ensures the very large levels have multiple exits so guards don't have to travel as far to respond to things inside/outside.